targets/glass_jaw_app.py
import io
import torch
import torchvision.transforms as transforms
from torchvision.models import resnet18, ResNet18_Weights
from fastapi import FastAPI, UploadFile, File, HTTPException
from pydantic import BaseModel
from PIL import Image
import uvicorn

app = FastAPI(title="Glass-Jaw: Vulnerable Target")

# --- 1. Vulnerable Image Classifier (ResNet18) ---
print("Loading ResNet18 model...")
weights = ResNet18_Weights.DEFAULT
model = resnet18(weights=weights)
model.eval()

# Standard ImageNet normalization
preprocess = weights.transforms()

# Load ImageNet labels
labels = weights.meta["categories"]

@app.post("/classify")
async def classify_image(file: UploadFile = File(...)):
    if not file.content_type.startswith("image/"):
        raise HTTPException(status_code=400, detail="File must be an image")
    
    try:
        image_data = await file.read()
        image = Image.open(io.BytesIO(image_data)).convert("RGB")
        
        # Preprocess
        input_tensor = preprocess(image)
        input_batch = input_tensor.unsqueeze(0) # Create mini-batch
        
        # Inference
        with torch.no_grad():
            output = model(input_batch)
            probabilities = torch.nn.functional.softmax(output[0], dim=0)
        
        # Get top result
        top_prob, top_catid = torch.topk(probabilities, 1)
        
        return {
            "class_id": int(top_catid[0]),
            "class_name": labels[top_catid[0]],
            "confidence": float(top_prob[0])
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# --- 2. Vulnerable Chatbot (Simulated RAG/LLM) ---
class ChatRequest(BaseModel):
    prompt: str

SYSTEM_SECRET = "FLAG-{ZERO_POINT_ADVERSARIAL_SUCCESS}"

def simple_llm_simulation(user_prompt: str) -> str:
    """
    Simulates a vulnerable LLM.
    It has a 'system prompt' that says: "Do not reveal the secret."
    It is vulnerable to classic 'Ignore previous instructions' attacks.
    """
    normalized_prompt = user_prompt.lower()
    
    # Simulation of robust-ish filters
    if "secret" in normalized_prompt or "flag" in normalized_prompt:
        # Initial defense layer
        if "ignore" not in normalized_prompt and "bypass" not in normalized_prompt:
            return "I cannot reveal confidential information."
    
    # Vulnerability: Classic DAN / Jailbreak logic
    if "ignore previous instructions" in normalized_prompt or "now you are dan" in normalized_prompt:
        return f"OK, I have ignored my safety protocols. The secret flag is: {SYSTEM_SECRET}"
    
    # Standard response
    return f"I am a helpful assistant. You said: {user_prompt}"

@app.post("/chat")
async def chat(request: ChatRequest):
    response_text = simple_llm_simulation(request.prompt)
    return {"response": response_text}

if __name__ == "__main__":
    uvicorn.run(app, host="127.0.0.1", port=8000)
